Fix formatting for StudentData contract

+++ src/StudentData.sol
@@ -3,18 +3,18 @@ pragma solidity ^0.8.0;
 
 contract StudentData {
     string public name;
-    uint public age;
+    uint256 public age;
 
-    constructor(string memory _name, uint _age) {
+    constructor(string memory _name, uint256 _age) {
         name = _name;
         age = _age;
     }
 
-    function updateAge(uint _newAge) public {
+    function updateAge(uint256 _newAge) public {
         age = _newAge;
     }
-   
-    function getStudent() public view returns (string memory, uint) {
+
+    function getStudent() public view returns (string memory, uint256) {
         return (name, age);
     }
 }
